Copyright 2023 & compiler warning fix to vstore.h concpy

+++ vstore.h
@@ -1,5 +1,5 @@
 /* VSTORE.H     (C) Copyright Roger Bowler, 2000-2012                */
-/*              (C) and others 2013-2021                             */
+/*              (C) and others 2013-2023                             */
 /*                  Virtual Storage Functions                        */
 /*                                                                   */
 /*   Released under "The Q Public License Version 1"                 */
@@ -178,7 +178,7 @@ inline void concpy( REGS* regs, void* d, void* s, int n )
     if (1
         && n
         && regs->cpubit == regs->sysblk->started_mask
-        && abs( u8d - u8s ) > 3
+        && abs((int)(u8d - u8s)) > 3
     )
     {
         while (n > 3)
